Fix Unlock for read-lock usecase
Read-lock can be held by multiple jobs concurrently. We need to make sure
that unlocking is not prone to race conditions in the read-lock scenario.

Instead of installing Lockable.this into DKV as a result of the Unlock
operation we are modifying the existing DKV instance ("old"). This removes
any side-effects from this atomic operation and its effects can be easily
rolled back.